pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mi_hgnn"
version = "0.1.0"
dependencies = [
  "numpy==1.26.4",
  "matplotlib==3.8.3",
  "networkx==3.2.1",
  "urchin==0.0.27",
  "torch==2.2.1",
  "torchvision==0.17.1",
  "torchmetrics==1.3.2",
  "torch_geometric==2.5.0",
  "pandas==2.2.1",
  "rosbags==0.9.22",
  "lightning==2.2.1",
  "wandb==0.16.6",
  "scipy==1.12.0",
  "names==0.3.0",
  "scikit-learn==1.4.1.post1",
  "PyYAML==5.4.1"
]
authors = [
  { name="Daniel Butterfield", email="dbutterfield3@gatech.edu" },
  { name="Jackson Cheng", email="ncheng49@gatech.edu" },
  { name="Lu Gan", email="lgan@gatech.edu" },
]
description = "A Morphological-Informed Heterogeneous graph neural network implementation that learns to estimate contact state on a robotic quadruped."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
]

[project.urls]
Homepage = "https://github.com/lunarlab-gatech/state-estimation-gnn"
Issues = "https://github.com/lunarlab-gatech/state-estimation-gnn/issues"